Move some files, hopefully increase coverage
