+++ bips/src/main/java/io/github/novacrypto/mnemonics/NumericEntryModel.kt
@@ -21,14 +21,14 @@
 
 package io.github.novacrypto.mnemonics
 
-import io.github.novacrypto.bip39.WordList
-
-internal fun WordList.toIterable() = object : Iterable<String> {
-    override fun iterator(): Iterator<String> {
-        var i = 0
-        return object : Iterator<String> {
-            override fun hasNext() = i < 2048
-            override fun next() = this@toIterable.getWord(i++)
-        }
-    }
+data class NumericEntryModel(
+        val currentKey: String,
+        val available: Set<Int>,
+        val exactMatches: List<String>,
+        val mnemonic: List<String>,
+        val display: String,
+        val previousState: NumericEntryModel? = null
+) {
+    val isBackSpaceAvailable: Boolean = previousState != null
+    fun isAvailable(buttonNumber: Int) = available.contains(buttonNumber)
 }
